Move listEmails tests to dedicated file to satisfy file_length lint (#124)
The new paging tests pushed ButtondownClientTests.swift past the 225-line
`file_length` limit. Relocate them into a focused
`ButtondownClientListEmailsTests.swift` suite (its own fixture/makeClient
helpers), restoring the original test file unchanged. swiftlint and
swift-format both report clean; all 12 tests pass.
