+++ Packages/BrightDigit/ButtondownKit/Tests/ButtondownKitTests/ButtondownClientTests.swift
@@ -198,64 +198,4 @@ import Testing
       _ = try ButtondownClient(apiKey: Self.apiKey)
     }
   #endif
-
-  /// `listEmails` reads a single page via `GET /emails`, decoding the
-  /// `EmailPage` (results + total count).
-  @Test internal func listEmailsSinglePage() async throws {
-    let transport = MockTransport(responses: [
-      "GET /emails": [.init(status: 200, json: try fixture("emails-page-1"))]
-    ])
-    let client = try makeClient(transport)
-
-    let page = try await client.listEmails(status: [.sent], page: 1)
-
-    #expect(page.count == 3)
-    #expect(page.results.count == 2)
-    #expect(page.results.first?.subject == "Issue One")
-
-    let recorded = await transport.recorded
-    let request = try #require(recorded.first)
-    #expect(request.method == "GET")
-    #expect(request.path.hasPrefix("/emails"))
-    // Status filter is applied server-side via the query string.
-    #expect(request.path.contains("status=sent"))
-    #expect(request.path.contains("page=1"))
-  }
-
-  /// `listAllEmails` walks every page (1-based) until the reported `count`
-  /// is reached, concatenating results in order — the paging contract.
-  @Test internal func listAllEmailsWalksPages() async throws {
-    let transport = MockTransport(responses: [
-      "GET /emails": [
-        .init(status: 200, json: try fixture("emails-page-1")),
-        .init(status: 200, json: try fixture("emails-page-2"))
-      ]
-    ])
-    let client = try makeClient(transport)
-
-    let emails = try await client.listAllEmails(status: [.sent])
-
-    #expect(emails.count == 3)
-    #expect(emails.map(\.subject) == ["Issue One", "Issue Two", "Issue Three"])
-
-    // Two page requests were issued; the second asked for page 2.
-    let recorded = await transport.recorded
-    #expect(recorded.count == 2)
-    #expect(recorded.first?.path.contains("page=1") == true)
-    #expect(recorded.last?.path.contains("page=2") == true)
-  }
-
-  /// A documented error status on `listEmails` surfaces as `unexpectedResponse`.
-  @Test internal func listEmailsUnexpectedStatusThrows() async throws {
-    let transport = MockTransport(responses: [
-      "GET /emails": [
-        .init(status: 403, json: #"{"detail":"nope","code":"forbidden"}"#)
-      ]
-    ])
-    let client = try makeClient(transport)
-
-    await #expect(throws: ButtondownClient.ClientError.unexpectedResponse) {
-      _ = try await client.listEmails()
-    }
-  }
 }
